chore(live-feedback-prompts): moved to its own feature folder
