+++ app/client/src/features/events/LiveFeedbackPrompts/LiveFeedbackPromptResponse/VoteResponseChart.tsx
