+++ app/client/src/features/events/LiveFeedbackPrompts/LiveFeedbackPromptResponses/index.ts
@@ -1,0 +1,4 @@
+export * from './LiveFeedbackPromptResponseList';
+export * from './PromptResponseAuthorCardHeader';
+export * from './ShareFeedbackPromptResults';
+export * from './useLiveFeedbackPromptResponses';
