--- app/client/src/features/events/Moderation/LiveFeedbackPromptResponses/index.ts
@@ -1,2 +1,0 @@
-export * from './useLiveFeedbackPromptResponses';
-export * from './ShareFeedbackPromptResults';
